no message
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@549 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ src/mcrypt/sha1.c
@@ -355,7 +355,7 @@ void gf_sha1_hmac( u8 *key, u32 keylen,
     memset( k_ipad, 0x36, 64 );
     memset( k_opad, 0x5C, 64 );
 
-    for( i = 0; i < keylen; i++ )
+    for( i = 0; i < (s32) keylen; i++ )
     {
         if( i >= 64 ) break;
